[refactor] common module package 구조 변경
