--- common/src/main/java/org/flab/deliveryplatform/common/exception/DeliveryPlatformErrorCode.java
@@ -1,8 +1,0 @@
-package org.flab.deliveryplatform.common.exception;
-
-public interface DeliveryPlatformErrorCode {
-
-    String name();
-
-    String getMessage();
-}
