+++ common/src/main/java/org/flab/deliveryplatform/common/web/exception/DeliveryPlatformErrorCode.java
@@ -1,11 +1,11 @@
-package org.flab.deliveryplatform.common.exception;
+package org.flab.deliveryplatform.common.web.exception;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
 @Getter
 @AllArgsConstructor
-public enum DeliveryPlatformDefaultErrorCode implements DeliveryPlatformErrorCode {
+public enum DeliveryPlatformErrorCode {
     BAD_REQUEST("잘못된 요청입니다.");
 
     private String message;
