+++ member-api/member-interfaces/src/test/java/org/flab/deliveryplatform/interfaces/member/web/WithdrawMemberControllerTest.java
@@ -9,8 +9,8 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import java.nio.charset.StandardCharsets;
 import java.util.UUID;
 import org.assertj.core.api.Assertions;
-import org.flab.deliveryplatform.common.web.DeliveryPlatformErrorResponse;
-import org.flab.deliveryplatform.common.web.DeliveryPlatformErrorResponse.DeliveryPlatformErrorResult;
+import org.flab.deliveryplatform.common.web.dto.DeliveryPlatformErrorResponse;
+import org.flab.deliveryplatform.common.web.dto.DeliveryPlatformErrorResponse.DeliveryPlatformErrorResult;
 import org.flab.deliveryplatform.interfaces.member.TestConfig;
 import org.flab.deliveryplatform.interfaces.member.web.exception.MemberErrorCode;
 import org.flab.deliveryplatform.member.application.port.SignUpMemberUseCase;
